LayoutTests/fast/dom/Geolocation/error-expected.txt

Tests Geolocation error callback using the mock service.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.POSITION_UNAVAILABLE
PASS error.message is mockMessage
PASS error.UNKNOWN_ERROR is undefined
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3
PASS successfullyParsed is true

TEST COMPLETE
